devtools/requirements.txt
click == 6.6
six == 1.10.0
protobuf == 3.0.0b3
grpcio
